+++ service/src/main/java/bio/terra/landingzone/stairway/flight/create/resource/step/CreateStorageAccountStep.java
@@ -2,12 +2,16 @@ package bio.terra.landingzone.stairway.flight.create.resource.step;
 
 import bio.terra.landingzone.library.configuration.LandingZoneAzureConfiguration;
 import bio.terra.landingzone.library.landingzones.definition.ResourceNameGenerator;
+import bio.terra.landingzone.library.landingzones.deployment.LandingZoneTagKeys;
+import bio.terra.landingzone.library.landingzones.deployment.ResourcePurpose;
+import bio.terra.landingzone.service.landingzone.azure.model.LandingZoneResource;
 import bio.terra.landingzone.stairway.flight.LandingZoneFlightMapKeys;
 import bio.terra.stairway.FlightContext;
 import bio.terra.stairway.StepResult;
 import bio.terra.stairway.StepStatus;
 import bio.terra.stairway.exception.RetryException;
 import com.azure.core.management.exception.ManagementException;
+import java.util.Map;
 import org.apache.commons.lang3.StringUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -15,6 +19,7 @@ import org.slf4j.LoggerFactory;
 public class CreateStorageAccountStep extends BaseResourceCreateStep {
   private static final Logger logger = LoggerFactory.getLogger(CreateStorageAccountStep.class);
   public static final String STORAGE_ACCOUNT_ID = "STORAGE_ACCOUNT_ID";
+  public static final String STORAGE_ACCOUNT_RESOURCE_KEY = "STORAGE_ACCOUNT";
 
   public CreateStorageAccountStep(
       LandingZoneAzureConfiguration landingZoneAzureConfiguration,
@@ -35,12 +40,29 @@ public class CreateStorageAccountStep extends BaseResourceCreateStep {
               .define(storageAccountName)
               .withRegion(resourceGroup.region())
               .withExistingResourceGroup(resourceGroup)
+              .withTags(
+                  Map.of(
+                      LandingZoneTagKeys.LANDING_ZONE_ID.toString(),
+                      landingZoneId.toString(),
+                      LandingZoneTagKeys.LANDING_ZONE_PURPOSE.toString(),
+                      ResourcePurpose.SHARED_RESOURCE.toString()))
               .create();
 
       context.getWorkingMap().put(STORAGE_ACCOUNT_ID, storage.id());
       context
           .getWorkingMap()
           .put(LandingZoneFlightMapKeys.STORAGE_ACCOUNT_NAME, storageAccountName);
+      context
+          .getWorkingMap()
+          .put(
+              STORAGE_ACCOUNT_RESOURCE_KEY,
+              LandingZoneResource.builder()
+                  .resourceId(storage.id())
+                  .resourceType(storage.type())
+                  .tags(storage.tags())
+                  .region(storage.regionName())
+                  .resourceName(storage.name())
+                  .build());
     } catch (ManagementException e) {
       if (StringUtils.equalsIgnoreCase(e.getValue().getCode(), "conflict")) {
         logger.info(
@@ -58,11 +80,16 @@ public class CreateStorageAccountStep extends BaseResourceCreateStep {
     // rollback here or in case of sub-flight do it there
     var storageAccountId = context.getWorkingMap().get(STORAGE_ACCOUNT_ID, String.class);
     try {
-      armManagers.azureResourceManager().storageAccounts().deleteById(storageAccountId);
+      if (storageAccountId != null) {
+        armManagers.azureResourceManager().storageAccounts().deleteById(storageAccountId);
+      }
     } catch (ManagementException e) {
       if (StringUtils.equalsIgnoreCase(e.getValue().getCode(), "ResourceNotFound")) {
+        logger.error(
+            "Storage account doesn't exist or has been already deleted. Id={}", storageAccountId);
         return StepResult.getStepResultSuccess();
       }
+      logger.error("Failed attempt to delete storage account. Id={}", storageAccountId);
       return new StepResult(StepStatus.STEP_RESULT_FAILURE_RETRY, e);
     }
     return StepResult.getStepResultSuccess();
