+++ service/src/main/java/bio/terra/landingzone/stairway/flight/create/resource/step/CreateVnetStep.java
@@ -2,12 +2,16 @@ package bio.terra.landingzone.stairway.flight.create.resource.step;
 
 import bio.terra.landingzone.library.configuration.LandingZoneAzureConfiguration;
 import bio.terra.landingzone.library.landingzones.definition.ResourceNameGenerator;
+import bio.terra.landingzone.library.landingzones.definition.factories.CromwellBaseResourcesFactory;
+import bio.terra.landingzone.library.landingzones.deployment.LandingZoneTagKeys;
+import bio.terra.landingzone.library.landingzones.deployment.SubnetResourcePurpose;
 import bio.terra.landingzone.service.landingzone.azure.model.LandingZoneResource;
 import bio.terra.stairway.FlightContext;
 import bio.terra.stairway.StepResult;
 import bio.terra.stairway.StepStatus;
 import bio.terra.stairway.exception.RetryException;
 import com.azure.core.management.exception.ManagementException;
+import java.util.Map;
 import org.apache.commons.lang3.StringUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -30,52 +34,58 @@ public class CreateVnetStep extends BaseResourceCreateStep {
     // Most like we need the same setup for different steps. At least we need armManagers.
     String vNetName = resourceNameGenerator.nextName(ResourceNameGenerator.MAX_VNET_NAME_LENGTH);
     try {
-      //      var vNet =
-      //          armManagers
-      //              .azureResourceManager()
-      //              .networks()
-      //              .define(vNetName)
-      //              .withRegion(resourceGroup.region())
-      //              .withExistingResourceGroup(resourceGroup)
-      //              .withAddressSpace(
-      //                  parametersResolver.getValue(
-      //
-      // CromwellBaseResourcesFactory.ParametersNames.VNET_ADDRESS_SPACE.name()))
-      //              .withSubnet(
-      //                  CromwellBaseResourcesFactory.Subnet.AKS_SUBNET.name(),
-      //                  parametersResolver.getValue(
-      //                      CromwellBaseResourcesFactory.Subnet.AKS_SUBNET.name()))
-      //              .withSubnet(
-      //                  CromwellBaseResourcesFactory.Subnet.BATCH_SUBNET.name(),
-      //                  parametersResolver.getValue(
-      //                      CromwellBaseResourcesFactory.Subnet.BATCH_SUBNET.name()))
-      //              .withSubnet(
-      //                  CromwellBaseResourcesFactory.Subnet.POSTGRESQL_SUBNET.name(),
-      //                  parametersResolver.getValue(
-      //                      CromwellBaseResourcesFactory.Subnet.POSTGRESQL_SUBNET.name()))
-      //              .withSubnet(
-      //                  CromwellBaseResourcesFactory.Subnet.COMPUTE_SUBNET.name(),
-      //                  parametersResolver.getValue(
-      //                      CromwellBaseResourcesFactory.Subnet.COMPUTE_SUBNET.name()))
-      //              .withTags(
-      //                  Map.of(
-      //                      CromwellBaseResourcesFactory.Subnet.AKS_SUBNET.name(),
-      //                      SubnetResourcePurpose.AKS_NODE_POOL_SUBNET.toString(),
-      //                      CromwellBaseResourcesFactory.Subnet.BATCH_SUBNET.name(),
-      //                      SubnetResourcePurpose.WORKSPACE_BATCH_SUBNET.toString(),
-      //                      CromwellBaseResourcesFactory.Subnet.POSTGRESQL_SUBNET.name(),
-      //                      SubnetResourcePurpose.POSTGRESQL_SUBNET.toString(),
-      //                      CromwellBaseResourcesFactory.Subnet.COMPUTE_SUBNET.name(),
-      //                      SubnetResourcePurpose.WORKSPACE_COMPUTE_SUBNET.toString()))
-      //              .create();
+      var vNet =
+          armManagers
+              .azureResourceManager()
+              .networks()
+              .define(vNetName)
+              .withRegion(resourceGroup.region())
+              .withExistingResourceGroup(resourceGroup)
+              .withAddressSpace(
+                  parametersResolver.getValue(
+                      CromwellBaseResourcesFactory.ParametersNames.VNET_ADDRESS_SPACE.name()))
+              .withSubnet(
+                  CromwellBaseResourcesFactory.Subnet.AKS_SUBNET.name(),
+                  parametersResolver.getValue(
+                      CromwellBaseResourcesFactory.Subnet.AKS_SUBNET.name()))
+              .withSubnet(
+                  CromwellBaseResourcesFactory.Subnet.BATCH_SUBNET.name(),
+                  parametersResolver.getValue(
+                      CromwellBaseResourcesFactory.Subnet.BATCH_SUBNET.name()))
+              .withSubnet(
+                  CromwellBaseResourcesFactory.Subnet.POSTGRESQL_SUBNET.name(),
+                  parametersResolver.getValue(
+                      CromwellBaseResourcesFactory.Subnet.POSTGRESQL_SUBNET.name()))
+              .withSubnet(
+                  CromwellBaseResourcesFactory.Subnet.COMPUTE_SUBNET.name(),
+                  parametersResolver.getValue(
+                      CromwellBaseResourcesFactory.Subnet.COMPUTE_SUBNET.name()))
+              .withTags(
+                  Map.of(
+                      LandingZoneTagKeys.LANDING_ZONE_ID.toString(),
+                      landingZoneId.toString(),
+                      CromwellBaseResourcesFactory.Subnet.AKS_SUBNET.name(),
+                      SubnetResourcePurpose.AKS_NODE_POOL_SUBNET.toString(),
+                      CromwellBaseResourcesFactory.Subnet.BATCH_SUBNET.name(),
+                      SubnetResourcePurpose.WORKSPACE_BATCH_SUBNET.toString(),
+                      CromwellBaseResourcesFactory.Subnet.POSTGRESQL_SUBNET.name(),
+                      SubnetResourcePurpose.POSTGRESQL_SUBNET.toString(),
+                      CromwellBaseResourcesFactory.Subnet.COMPUTE_SUBNET.name(),
+                      SubnetResourcePurpose.WORKSPACE_COMPUTE_SUBNET.toString()))
+              .create();
 
-      context.getWorkingMap().put(VNET_ID, "vNet.id()");
-      var vNetId = "vNetTestId";
+      context.getWorkingMap().put(VNET_ID, vNet.id());
       context
           .getWorkingMap()
           .put(
               VNET_RESOURCE_KEY,
-              LandingZoneResource.builder().resourceId(vNetId /*vNet.id()*/).build());
+              LandingZoneResource.builder()
+                  .resourceId(vNet.id())
+                  .resourceType(vNet.type())
+                  .tags(vNet.tags())
+                  .region(vNet.regionName())
+                  .resourceName(vNet.name())
+                  .build());
     } catch (ManagementException e) {
       if (StringUtils.equalsIgnoreCase(e.getValue().getCode(), "conflict")) {
         logger.info(RESOURCE_ALREADY_EXISTS, "Virtual network", vNetName, resourceGroup.name());
@@ -89,15 +99,18 @@ public class CreateVnetStep extends BaseResourceCreateStep {
 
   @Override
   public StepResult undoStep(FlightContext context) throws InterruptedException {
-    var vNetId =
-        context.getWorkingMap().get(VNET_RESOURCE_KEY, LandingZoneResource.class).resourceId();
+    var vNetId = context.getWorkingMap().get(VNET_ID, String.class);
     try {
-      // armManagers.azureResourceManager().networks().deleteById(vNetId);
-      armManagers.azureResourceManager().genericResources().deleteById(vNetId);
+      if (vNetId != null) {
+        // armManagers.azureResourceManager().networks().deleteById(vNetId);
+        armManagers.azureResourceManager().genericResources().deleteById(vNetId);
+      }
     } catch (ManagementException e) {
       if (StringUtils.equalsIgnoreCase(e.getValue().getCode(), "ResourceNotFound")) {
+        logger.error("Virtual network doesn't exist or has been already deleted. Id={}", vNetId);
         return StepResult.getStepResultSuccess();
       }
+      logger.error("Failed attempt to delete virtual network. Id={}", vNetId);
       return new StepResult(StepStatus.STEP_RESULT_FAILURE_RETRY, e);
     }
     return StepResult.getStepResultSuccess();
